+++ Packages/BrandonUtils/Runtime/Saving/SaveDataTestImpl.cs
@@ -8,6 +8,8 @@
     public class SaveDataTestImpl : SaveData<SaveDataTestImpl> {
         public string Word;
 
+        public string Word2 = "My Mom";
+
         public SaveDataTestImpl(string nickname) : base(nickname) { }
     }
 }