+++ Packages/BrandonUtils/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.0] - 6/29/2021
+
+### Added
+
+- Generic `ISaveData<T>` interface
+- Ability to reset `SaveData` values using some fancy `JsonConvert` shenanigans
+
+### Changed
+
+- `ReflectionUtils.Construct()` can now find private constructors
+- `ISaveData` timestamps are now `DateTime?`
+- `SaveData` no longer initializes timestamps
+
 ## [2.4.0] - 6/29/2021
 
 ### Added
@@ -44,7 +57,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
-- `SaveData.Reload()` become `virtual`
+- `SaveData.Reload()` became `virtual`
 
 ## [2.0.0] - 6/26/2021